access to read_from and write_out changend
also added read_from and write_out to dense matrix

For all matrix data container read_from_*FileMode*(string filename) and write_out_*FileMode*(string filename) were deleted.
Instead read_from(FileMode mode, string filename) or write_out(FileMode mode, string filename) are to be used.
Thereby FileMode is an enum of the possible datatypes. For further information please refer to the LAFEM design documentation.
For usable FileMode types, please refer to the corrosponding header file.
A quick userguide:
Say you would like to save a dense_vector dvec in mtx format to a file with its path in string test.
Then the call
        dvec.write_out(FileMode::fm_mtx, test);
will do the trick.
For reading in a dense_vector to dvec from file test in mtx format, use
        dvec.read_from(FileMode::fm_mtx, test);
.
In general its always possible to use FileMode::fm_binary to write or read in a binary file of the same container type.
If a FileMode is not useable, an error message will be printed out.
Warning: It will not be checked, if the used FileMode is compatible with the file that is to be read in.