Fix documentation and adjust style
I just wanted to do some no brainer stuff so I did the following:
- Fixed some documentation errors.
- Changed VT_ to ValueType when it's a TinyVector and not a "big" one.
  (otherwise "Common typename abbreviations" in "LAFEM design
  description" would have been misleading.)
- Changed const references to primitive types as return values to only
  the primitive type.
- Changed the zero_element() function of the blocked SparseVector to
  match the same function of the blocked CSR matrix.
  Two reasons for this way are: The conversion method BCSR -> CSR relies
  on zero_element() of BCSR to return a DataType. Also there are no
  binary == and != operators for TinyMatrix.
- Some other minor style stuff.

Resulting from the changes above here is a question to discuss:
As now all `zero_element()` implementations effectively return
`DataType(0)` is that function really (still) useful?
If not then `_scalar_dt[0]` in which the value of the zero element is
stored becomes obsolete.