Added utility to pack and unpack with zfp library.
Zfp is a lossy compression bib, which uses continuity in the given data sets to compress data with a maximum error to be set.
For continuous solutions of a solver this compression is significantly better for big data sets compared to lossless algorithms like ZLIB.
For best results the data should ordered by continuity, so the data should only have small changes while going through the data array.

A quick user guide:
It is recommended to use zfp through the encode and decode functions in pack.hpp.
The used compression mode solely depends on the given parameter Pack::Type.
For zfp compression use PF32 or PF64.
The function encode expects a buffer to write the compressed data into. To guarantee these buffer is sufficient in size, use the estimate_size function with the same parameter you want to use for encoding.
Besides the standard parameters as seen in the header file, giving a tolerance which should not be exceeded is highly recommended, as the compression rate depends on the maximum error.
As principle: The higher the tolerance, the better the compression.

For decoding, again use Pack::Type to use the zfp decompression.
Important: ZFP compressed data should only be decompressed with zfp...