added VC16 support for ZFP library
Note that for the current stable version 0.5.5 of the zfp library, it is
necessary to patch the source files by using the vc_patch_zfp.py script,
because it is not possible to link it as a static library otherwise.
However, the zfp code that is hosted on GitHub is already fixed, so the
patch script will most probably become obsolete with the next stable
release of the zfp library.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>