implemented XABORTM macro for program abortion
This new macro is an addition to the [X]ASSERT[M] macro family,
which is effectively equivalent to an unconditionally failing
assertion and therefore always aborts program execution.

This new macro as well as its underlying implementation of
the new FEAT::abortion() function are replacing the old
'throw InternalError(...)' statements, which have not been
replaced before because they were triggered by an if-statement
or some other means which could not be emplaced as an asserted
expression in an XASSERTM call.

In consequence, any occurance of

  throw InternalError(__func__, __FILE__, __LINE__, msg);

has now been replaced by

  XABORTM(msg);

or by an appropriate

  XASSERTM(condition, msg);

where 'msg' is some null-terminated C string or a std::string
containing the error message and, where applicable, 'condition'
is the negated condition which triggered the previous throw.

It is recommended to use the XABORTM and [X]ASSERT[M] macros
in future instead of throwing an InternalError exception.

This commit also introduces an overload to the FEAT::assertion()
function which accepts an error message in the form of a
std::string object rather than a raw C-string.
This is useful when composing error message of multiple parts.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>