implemented DistFileIO::write/read_combined functions
This new pair of distributed I/O functions is a slightly more
sophisticated version of the existing write/read_ordered functions,
which only offered a 'raw' I/O functionality without any included
possibility of storing/reading buffer sizes.

The new functions write out/read in a combined binary file, which
starts with a binary header that stores the number of processes used
to create the file as well as a table of all data array sizes.
In addition, the file can also contain a 'common' data darray,
which is broadcasted to all processes upon reading and which can
be used to store data that is identical for all processes and
therefore does not need to be stored for each process individually
to reduce the total file size.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>