Added zfp compression to the checkpointing system
This update implements changes to the serialise method found in all containers
in /kernel/lafem and the CheckpoitnControl class in /control.

Serialise() has now a new parameter SerialConfig. This configuration class in
/kernel/lafem/container holds the compression parameters, which define the
use of the compression libraries in serialise().
There are two recommended ways to handle compression.
First, just use the constructor without input parameters, which sets lossless
compression if ZLIB is loaded, but never lossy compression.
Second, to have more control over the used compression, use the constructor
SerialConfig(bool zlib_compression, bool zfp_compression, FEAT::Real tol = FEAT::Real(-1)).
This constructor gives you the choice whether to use ZLIB compression, which
enables lossless compression for the _elements and _indices arrays found in all
containers, and/or to use ZFP compression, which enables lossy compression for
the _elements array. If ZFP compression is enabled, tol has to be a value
strictly greater zero, setting a maximum error for the compressed values.
If both parameters are true, _elements will be compressed by ZFP, _indices by ZLIB.

A note to the write_out method for container classes:
write_out() will always use the default compression defined in the default
constructor of SerialConfig.

The other change implemented is that SerialConfig is also used to handle the
compression in CheckpointControl, while from now on only one file will be
written out.
To choose the used compression just create a SerialConfig object with the wished
compression mode and pass it on to the constructor of CheckpointControl.
Alternatively just use the default parameter.

An important note: Code that uses the CheckpointControl class will run
without any adjustment. But due to changes of the write method, pre change
files can not be read in anymore.