added ParsedVectorFunction and renamed ParsedFunction to ParsedScalarFunction
This commit adds a vector-valued equivalent implementation 'ParsedVectorFunction'
of the ParsedFunction class, which now has been renamed to 'ParsedScalarFunction'
for the sake of consistency. Additionally, both implementations have been now
can manage a set of user-defined variables, which can be added and modified by
the 'add_variable' and 'set_variable' functions.

Moreover, this commit adds a simple test for the ParsedScalarFunction and
ParsedVectorFunction classes.

Note: For the sake of downward compatibility, a template alias named
ParsedFunction has been added. It is recommended to switch to the class name
ParsedScalarFunction after rebasing.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>