fix copy/paste error in the cellwise evaluation method of RumpfFunctionals
The error led to false functional values in the case of 3D RumpfFunctionals
because the cofactor part was not correct.
In the case of 2D meshes the cofactor part is not needed, so all 2D tests and
applications were correct. Furthermore, the eval_fval_grad() method which
computes the functional value and its gradient was correct.
It seems like the HyperelasticityFunctional-test was the only place where the
wrong method was used in a 3D context to evaluate the functional value before
and after the optimizations process.
That the previous implementation was wrong can be verified by comparing the
cellwise evaluation function to the unrolled 3d-p1 Rumpffunctional or by
comparing the results of the eval_fval_cellwise() and eval_fval_grad()
methods.

Signed-off-by: Jonas Duennebacke <jduenneb@math.tu-dortmund.de>