avoid division-by-zero for singular grad(R) matrix in Hyperelasticity stuff
The code worked just fine even before this patch, because the outer solver
handled the singular matrix case appropriately after the divsion by zero happened.
However, some modern runtime sanitizers explicitly catch floating point
division-by-zero events -- which are not considered an error per se --
and emit runtime warnings, which then in turn are displayed in our nightly
regression test system as an error.

So, we now check the determinant before trying to invert the matrix.
Note that the determinant already had to be computed for some other
computations, so this should not result in a runtime overhead.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>