thirdparty download scripts can now patch the unpacked sources
The ThirdpartyPackage class has been enhanced with a 'patch' method
which is called after a source TAR/GZ/ZIP archive is unpacked.

This has become necessary to ensure that the HYPRE library can be
used in combination with the clang 10 (or higher) compiler, because
the compiler and the library had a slight disagreement on who had
the right to name a file 'version'.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>