Added parallel FETI implementation of poisson in MPI
This adds three cpp files, which only are buildable if
UMFPACK is available:
- area51/FETI_seq_usquare.cpp which implements a sequential FETI
implementation of the poisson-dirichlet problem on the unit square
and will be maintained due to testing purposes independent of MPI.

- area51/FETI_para_usquare.cpp which realises a MPI implementation
of FETI and uses UMFPACK as local and projector solver.
This application has no testing and timing capabilities and acts as
stepping stone to a multigrid implementation.

-area51/FETI_muligrid_usquare.cpp which realises the former with a
multigrid(smoother jacobi/coarse UMFPACK) as local solver.
This application has testing and timing capabilities.
Should only be called by 4, 16 or 64 processes due to mesh generator,
generally can handle multiples of 4 but an internal check has to be
adjusted.

TODO: Generalise FETI solver and embed into Solver framework.