defined FEAT_PRAGMA_OMP macro for OpenMP pragmas
This commit introduces the FEAT_PRAGMA_OMP macro, which is to be used for
declaring OpenMP pragmas. The reason why one should use this macro instead of
the usual plain '#pragma omp' directives is that the macro can be disabled by
the build system if FEAT is built without OpenMP support, in which case some
compilers would issue an 'unknown pragma' warning otherwise.

This macro is used by passing the pragma directive arguments as macro
arguments, e.g. the pragma directive

#pragma omp parallel for reduction(+:r) shared(alpha) if(n > 100)

becomes

FEAT_PRAGMA_OMP(parallel for reduction(+:r) shared(alpha) if(n > 100))

Also, for the sake of consistency, this commit also renames the 'FEAT_IVDEP'
macro to 'FEAT_PRAGMA_IVDEP'.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>