switched convention from British English to American English
This commit changes the convention language dialect from British English to the
more common American English. Note that the most prominent change is the change
of the 'ise' and 'isation' endings to 'ize' and 'ization'. However, since this
is a rather common ending in technical texts and source codes there are a lot
of files which are affected by this change, especially since every executable
in FEAT is required to call Runtime::initialize() and Runtime::finalize().
Except for these endings, the only additional changes is that 'colour' has been
replaced by 'color' and 'neighbour' has been replaced by 'neighbor', which also
resulted in a few renamed files.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>