minor overhaul of the thirdparty package system
This commit makes some changes to the directory/file naming pattern
used in the thirdparty library directory, which was a wild mixture
of different patterns before.

Now, every downloaded TAR/GZ/ZIP archive contains both the name of
the library as well as the version number, assuming that the library
does have a version number. Currently, the following libraries do
not have version numbers: deathhandler, triangle, floatx.

Consistently, the directories, which contain the unpacked sources,
are also named including the library name and version number.

Finally, most of the version number mumbo jumbo has been isolated
in the thirdparty python scripts, which also provide the directory
names of the thirdparty libraries to CMake via defines. This way,
we do not have to adjust the CMakeLists.txt files every time we
update to a new library version, unless there are some additional
errors or warnings or include paths to take care of.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>