Graph, Coloring and Permutation now use std::vector internally
The access from ouside is not changed.
I also added a new constructor to the class Graph and Coloring class
which gets handed a vector / vectors.
The reason for these changes is not to use bare arrays anymore and
instead to use std::vector.
std::vector has a size function, so we are able to eliminate some
member varriables.