use ExpBubbleFunction as analytical solution in Tutorial 01
The sine-bubble has been used in Tutorial 01 for about 8 years now
and still people are surprised by the fact that the unpreconditioned
CG solver converges after 1 iteration, because the sine-bubble is an
eigenfunction of the Laplace operator.

So now, the solution has been set to the ExpBubbleFunction, which
has a slightly more complicated definition and a significantly more
complicated right-hand-side function (which is not needed here).
The ExpBubbleFunction has been chosen instead of the more simple
Q2BubbleFunction, because the former one is non-polynomial and
therefore gives useful H0/H1-Errors even for higher order FEM spaces,
whereas the latter one gives H0/H1 errors equal to zero from Q2 onwards.
At least this should make all people happy who do not like
to deal with iterative solvers that pretend to be direct solvers.

Note that the solution is quite similar to the SineBubbleFunction
in the 'Eye-Norm' (i.e. looking at it), so you should see any huge
difference in the resulting VTK files. Also, the absolute H0/H1
errors as well as the CG iteration numbers are comparable to that
of the SineBubbleFunction.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>