implemented MeshPermutation and added clone method for Geometry classes
This commit adds the Geometry::MeshPermutation class, which is now used by the
ConformalMesh, MeshNode and PartiDomainControl classes to create and manage
a permutation of the mesh entities. In addition to that, the MeshPermutation
object can also manage an element coloring or element layering, which can be
used by multi-threaded assembly classes, which will be added soon.

The classes which are related to the assembly of inter-grid matrices, namely
the Assembly::SymbolicAssembler as well as the Assembly::GridTransfer classes,
have been adjusted accordingly to ensure that they can assemble the correct
matrices even if the coarse and/or fine mesh objects have been permuted.

On a side node: currently, mesh permutations are not exported to mesh files
when using the Geometry::MeshFileWriter, because the mesh file format does
not support this feature yet.

In addition to that, this commit also adds the infamous clone() method to the
most important Geometry classes, so that ConformalMesh, MeshPart and MeshNode
objects can now be cloned.

Furthermore, this commit adds some documentation additions and improvements
for various geometry classes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>