Implemented general form of Bernstein2 node functionals.
This is necessary, because the coefficient changes for every Element.
The previous implementation only worked for regualr shapes like squares,
but did not provide the correct interpolation for general shapes,
especially when isoparametric transformations were used.

Be aware, that because of this the application of the interpolation is computionaly expencive
and therefore one should avoid performing redundant interpolations if these can be avoided.