documented transient vs resident pointer/reference parameter properties
This commit introduces two new doxygen commands named \transient and \resident
which are used to document whether or not a (class member) function stores an
internal pointer/reference to a pointer/reference function parameter even after
the function call returns.

In short:
A pointer/reference function paramter is called transient, if the object
referenced by the pointer or reference is only accessed within the function body,
but no reference or pointer to the object is kept in the class object after the
function call returns.

A pointer/reference function parameter is called resident, if a reference or
pointer to the referenced object is stored inside a class object for later use
after the function call returns. In consequence, the pointed/referenced object
has to remain valid after the called function returns until the lifetime of the
resident object (usually the class object of the member function, unless stated
otherwise in the function documentation) reaches its end.

Also, any pointer/reference parameter that is not explicitly declared as either
resident or transient, is implicitly transient.

It is to be said that the classes in the LAFEM, Global and Solver namespaces
have not been documented with respect to the transient/resident property yet,
because Dirk and Lisa-Marie are heavily at work in these namespaces to convert
all classes to the new unified memory interface and thus re-documenting these
classes would lead to tons of rebase conflicts. In consequence, the classes in
these namespaces will be documented with respect to the transient/resident
property after the unified memory update has been rolled out.

On a side note, this commit also includes a quarter billion documentation fixes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>