successfully circumvented another one of Intel's opt-mode Heisenbugs(TM)
This time, the statement "coef = coef * (2 / mean)" did not convince the
Intel compiler that the actual value of 'mean' is kind of important for
the result of the program and so the ICC decided to opt away this variable
or even the whole scaling statement. Fortunately, "coef *= (2 / mean)"
seemed to have triggered some more motherly instincts and the statement
is now correctly executed. Furthermore, this statement is also a bit more
efficient, because it saves one temporary object creation and assignment.
I have also taken the freedom to add a proper typecast to DataType(2) for
the sake of pedancy and consistency.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>