various small updates for the TraceAssembler incl. normal vector computation
This commit adds the (automatic) computation of facet normal vectors for the
TraceAssembler. The normal vector is stored in the TrafoEvalData structure,
which is passed to the operator/functional's evaluator during the set_point()
function call as the 'tau' parameter, i.e. you can access it as 'tau.normal'.

Furthermore, this commit also includes a small refactoring of the TraceAssembler
class, most notably the 'compile_facets' function has been renamed to 'compile'
for the sake of consistency with the DomainAssembler class and the one and only
function parameter 'only_boundary' has been removed. Also, the documentation
of the TraceAssembler class has been enhanced significantly.

Moreover, this commit fixes a bug in the reference shape facet orientation
specification in the Shape::FaceTraits class for the Tetrahedron (Simplex<3>)
shape. Apparently, no one has used this facet orientation information before.

Finally, this commit also includes a bunch of small documentation fixes and
improvements.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>