use Assembly::DomainAssembler in applications
An instance of the DomainAssembler class has been added to each derived class
of the Control::DomainLevel class template, so now every level in a domain
control object contains a domain assembler that can be used by the application
for its assemblies, which could even be performed with multiple threads.

This commit adapts almost all applications to use the not-so-new-anymore
DomainAssembler class, although the applications do not make use of the multi
threading functionality yet.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>