fixed a serious bug in signed distance computation of Bezier spline class
The previous implementation contained a logical error in the computation of the sign in the signed
distance computation of the Bezier spline class, where the sign might have been incorrect for points
for which the closest candidate point on the Bezier spline is a vertex, whose two adjacent curve
segments form an angle of less than 90° or more than 270°. This is a rather rare case in most real
life examples in the FEM context, so this bug was apparently not discovered before.

The new implementation is based on the 'geom_polygon_projector' routine from FEAT2 that I had come
up with some ~15 years ago, which handles the case of a vertex candidate separately and determines
the correct sign by checking whether the two adjacent segments form a convex or concave angle.
Since you really enjoy reading these commit novels (and because it took me half an hour to get the
ASCII arts right), here comes an excerpt from the code comment:

Now we want to check whether the angle between the two curve segments is < 180 deg, = 180 deg
or > 180 deg and we do this by computing the normal vectors, then turning the first one by 90 deg,
computing the dot-product and checking that against 0. In the two mighty fine pieces of art below,
'X' is the common candidate vertex, 'Prev' is the previous segment, 'Next' is the next segment,
'T' is the tangent of 'Prev' at 'X' and 'N' is the (outer) normal vector of 'Next' at 'X'. By
computing the dot product of 'N' and 'T' we can find out whether the vertex is convex (N*T > 0,
left image) or concave (N*T < 0, right image). Note that the vertex 'X' can only be a candidate
for world points 'W' which lie in the concave area (outside on left image, inside on right image),
because if the world point 'W' were inside the convex area, we would have found a candidate point
on either 'Prev' or 'Next' that is closer to the world point than the vertex 'X', so we would not
be in this if-clause at all.

                     outside / inside               outside \ inside
                            /                      '-.       \
                      Prev /                         N '-.    \ Prev
            W             /                                '-. \
                        .X                                     'X      W
                    .-' ' \                                    / '
             N  .-'    '   \ Next                        Next /   '
            .-'       '     \                                /     ' T
                     ' T     \                              /       '

In short: if the candidate vertex 'X' is a convex vertex (N*T > 0), then the world point 'W'
lies on the outside of our Bezier figure and if the candidate vertex 'X' is a concave vertex
(N*T < 0) then the world point 'W' is on the inside of our Bezier figure. Note that if the
vertex is straight (N*T = 0), then we have already figured out the correct sign by checking
it against the normal of either segment before, i.e. 'best_sign' is already correct.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>