implemented class Geometry::MeshDistortion
This new class can be used to distort/disturb a conformal mesh by moving
the inner vertices of the mesh in a random manner. The class supports
three different types of distortion algorithms: the first one uses a
uniform scaling factor for all vertices, the second one automatically
chooses the scaling factor based on the shortest edge length uniformly,
and the third one uses the shortest edge length of each vertex locally.