Added joined solver operation for the unsteady interface and fixed tests.
This adds two additional joined solver calls, which calculate the solution to the extended
Navier-Stokes in parallel and gather the solution on the root process to a global solution.
For now, can only be called with multiple processes.
It is required to set the "--joined-sol" commandline parser argument so the root-mesh is assembled.
Also adds a joined_get_mesh method to retrieve the "global" root mesh.

Furthermore this fixes some namespace and solver status bugs. Also adds a cmake subdirectory and adjusts tests.
Additionally adds a get_partioned_mesh() method to retrieve the local mesh.