added "simple" CCND applications, including DirectStokesSolver class
This commit adds a family of "simple" steady and unsteady CCND applications,
which are preconfigured for a small set of simple benchmark problems. Please
refer to the readme.txt in applications/ccnd_simple for more information.

Furthermore, this commit also adds the Solver::DirectStokesSolver class template
which, in its current state, is a somewhat more sophisticated hybrid of the
GenericUmfpack and SaddleUmfpackMean solvers. This class converts a Stokes-like
saddle-point system of the form (where usually S=0)

  [ A  B ] * [ v ] = [ f ]
  [ D  S ]   [ p ] = [ g ]

along with its velocity and pressure filters into a CSR system of the form

  [ A   B   Q   0 ]   [ v ]   [ f ]
  [ D   S   0   P ] * [ p ] = [ g ]
  [ Q^T 0   0   0 ]   [ q ]   [ 0 ]
  [ 0   P^T 0   0 ]   [ r ]   [ 0 ]

where Q is the velocity Lagrange-multiplier matrix, which derives from the slip
filters, and P is the pressure Lagrange-multiplier matrix, which derives from
the pressure integral mean filter. Note that both Q and P may be empty matrices.

In its current implementation state, the DirectStokesSolver can only solve the
system on a single process, so it may be used as a single-process coarse grid
solver for a multigrid. Maybe this class will also be extended to assemble a
partitioned CSR system, so that parallel direct solver as SuperLU may be used
to solve the system in the future.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>