implemented base-splitter for joined vector I/O in parallel apps
This commit introduces a new class Global::Splitter, which can be used by
MPI parallel applications to join the partitioned solution vector into a single
vector which corresponds to the DOFs of the sequential (i.e. unpartitioned)
solution vector on the same refinement level. This joined solution vector can
then be read in by an(other) application running in serial or parallel with a
potentially different number of MPI processes and then split onto the new
partitioning by the splitter class, thus obtaining a partitioned solution vector.
Obviously, this also works in non-MPI builds and MPI-build applications running
with only one MPI process.

Note that there is an unavoidable drawback here: if you want to make use of this
new functionality in an application, then you cannot use more than a single
partitioning layer with more than 1 process, i.e. truly multi-layered partitioning
is not supported. However, this is not a real show-stopper, because you only
really need (true) multi-layered partitioning for MPI runs with > 1000 processes
anyways and in this case it is impossible to collect an entire serial vector on
a single process without running out of memory anyways.

It is to be said that the above "drawback" does not void the possibility to
write apps that support multi-layered partitionings, i.e. an application can
easily be implemented in a manner that it supports both multi-layered partitionings
as well as the base-splitter-approach for writing/reading joined vectors. The
app can even be compiled to support both features and it can be chosen at runtime
whether one wants to make use of multi-layered partitioning XOR joined vector I/O
via the base-splitter -- it is only impossible to use both features in the same
application run.

To simplify the usage, the system control classes Control::ScalarBasic and
Control::StokesBlocked (as well as their derived classes, of course) have been
enhanced with corresponding base splitter objects as well as the required
assembly functions.

To enable the base-splitter functionality in an application, one first has to
call the new 'keep_base_levels' function of the domain control before calling
its 'create' function and then call the 'assemble_base_splitter' function of
the finest system level. Then, one can make use of the Global::Splitters
functions to join/split vectors into a single local vector on the rank 0 process
or write out/read from file directly.

Currently, only the dfg95 ccnd apps make use of this new functionality.

To avoid confusion, an old class named Assembly::BaseSplitter has been removed,
because it was obsolete (and thus unused) since the time that we euthanized the
"Foundation" module of the FEAT kernel.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>