use defect-based volumetric drag/lift force computation in CCND
The previous version of the volumetric drag/lift force computation routines
were based on an implementation in FeatFlow2, which itself was based on some
implementation by someone using FeatFlow 1, however, no one really seems to
remember how exactly that variant was derived.

The new implementation uses the final defect vector to compute the forces via
a dot-product with the characteristic function of the obstacle boundary, which
yields drag/lift errors of the order O(h^4), which is a lot better than the
order of O(h^2), which is obtained by the line integration variant.

See the following paper for more details:
M. Giles, M.G. Larson, J.M. Levenstam, E. Süli: "Adaptive Error Control For
Finite Element Approximations Of The Lift And Drag Coefficients In Viscous Flow"

Note: The CCND solver is going to be completely rewritten from scratch within
the next few months, so this is just a quick "hotfix" to enable some master
student to utilize the new functionality in the master thesis implementations.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>