implemented vector-type convert function for Gate, Muxer and Splitter + doxy
The original convert function is only suitable if only the MemType, DataType or
IndexType are different but not if the actual VectorType container is different.
The new (additional) convert function accepts an already compiled Gate/Muxer/Splitter
for a different vector container type as well as a temporary vector object of
the new container as input.

This functionality can be used to easily convert a Gate/Muxer/Splitter object
working with DenseVector into a corresponding object working with DenseVectorBlocked
or vice versa.

Note that in the case of the Splitter, one still needs to provide a base-mesh
vector template separately if one wants to use the join or join_write_out
functions -- it is not required for the split or split_read_from functions, though.

Also heavily improved the documentation of the Gate, Muxer, Splitter, Matrix and
Vector classes of the Global namespace.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>