bring two major code updates to DFG95 CP2D application
This commit introduces a few important changes to the CP2D applications, which
brings the code slightly more up to date with FEAT's current features. The two
major changes are:

1. Enable the code to use recursive partitioning and therefore enable it to use
UMFPACK as a coarse grid solver for both the A as well as the S systems

2. Use the Global::PMDCDSCMatrix class instead of the Global::SymmetricLumpedSchurMatrix
to approximate the pressure Poisson problem, which yields a significant speedup.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>