added support for Sandia's Zoltan hypergraph partitioner
This commit commit adds a class that acts as a backend to the Zoltan hypergraph
partitioning suite from the Sandra National Laboratories, which is attached
to FEAT's PartiDomainControl and therefore allows the use of Zoltan as a
domain decomposition partitioner and can be seen as an alternative to the more
well known ParMETIS partitioner.

Just as ParMETIS, Zoltan is included as a third-party library and has to be
enabled by specifying the 'zoltan' tag as part of the build-id. Furthermore,
one also has to specify 'zoltan' as one of the tags passed to the '--parti-type'
command line option, because it is deactivated by default since for most of our
commonly used meshes the default 'naive' partitioner generates perfectly reasonable
decompositions at virtually no cost. However, once we get to import more complex
meshes, a proper partitioner as Zoltan or ParMETIS becomes essential.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>