added support for ParMETIS geometric K-way partitioner
This commit replaces the old METIS-based partitioner backend implementation by
a (more) proper ParMETIS backend, which now also utilizes the vertex coordinates
of the underlying mesh to improve partitioning quality. The implementation of
the backend is quite similar to that of the Zoltan backend and all header
includes of ParMETIS headers are limited to a single source file to avoid
namespace pollution.

Also added a check for both Zoltan and ParMETIS backends to ensure that a
partitioning is only accepted if all patches are non-empty, which unfortunately
is not always the case for both Zoltan and ParMETIS. In the case that a
partitioning is not accepted, the PartiDomainControl class will try the next
available partitioning algorithm to find an appropriate partitioning.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>