implemented Visual Studio 2022 build system
This commit adds the basic project files as well as a bunch of Python scripts,
which together yield the new Visual Studio 2022 based build system for Windows.

Just as with the previous Visual Studio 2019 based build system, project and
solution files are generated by a Python script named 'vc17-gen.py' located in
the FEAT root directory.

One major overhaul is the management of the third-party libraries, which is now
all performed by a single Python script named 'vc17-thirdparty.py' located in
the 'thirdparty' subdirectory. This script is row responsible for downloading,
unpacking, patching and compiling the third-party libraries that the user wants
to use. Also, the way the build system recognizes installed third-party libraries
has been revamped seriously, however, this is mostly an under-the-hood change
that no user should need to bother about.

One other major change is that the new build system does not support 32-bit
builds anymore, which should hopefully not be any problem for anyone who is not
running a museum computer collection.

Currently, the new build system does not support CUDA or MKL -- the support for
these backends will be added once the unified memory system has been employed.

The Visual Studio 2019 build system will be removed shortly after to avoid
having to maintain two slightly incompatible Windows build systems, especially
with the major CUDA update lurking on the horizon.

Signed-off-by: Peter Zajac <peter.zajac@math.tu-dortmund.de>