implemented analytic lambda expression function classes
This commit introduces a family of five closely related classes implementing the Analytic::Function
interface by allowing the user to specify function formulae by supplying them as standard C++ lambda
expressions:
- Analytic::LambdaScalarFunction1D : R^1 -> R
- Analytic::LambdaScalarFunction2D : R^2 -> R
- Analytic::LambdaScalarFunction3D : R^3 -> R
- Analytic::LambdaVectorFunction2D : R^2 -> R^2
- Analytic::LambdaVectorFunction3D : R^3 -> R^3

as well as their corresponding construction helper function overloads:
- Analytic::create_lambda_function_scalar_1d()
- Analytic::create_lambda_function_scalar_2d()
- Analytic::create_lambda_function_scalar_3d()
- Analytic::create_lambda_function_vector_2d()
- Analytic::create_lambda_function_vector_3d()

The purpose of these classes and create helper functions is to provide a more convenient way of
implementing custom simple analytic functions without the need to implement a class for this by
hand and without using the Analytic::ParsedScalar/VectorFunction implementations, which rely on
the fparser third-party library and which require parsing and evaluating the formulae at runtime,
which is a more or less serious negative runtime performance impact.

One may optionally also specify the formulae for the first and second partial derivatives as lambda
expressions in addition to the function values themselves. However, for the sake of convenience,
one may also omit the specification of the derivatives, in which case the Lambda function classes
will automatically enable a Richardson extrapolation scheme based on second-order central difference
quotients, which is then used to compute approximations of the derivatives that are then used to
provide function gradients and hessians.

The usage of these new classes is as easy as it can get with lambda expressions, e.g. for the simple
scalar 2D function x^2 - y^2 the corresponding creation of an analytic lambda function without
specifying any partial derivatives may look like:

auto saddle_func = create_lambda_function_scalar_2d( [](auto x, auto y){return x*x - y*y;} );

For more details and examples please consider the page "Analytic Lambda Expression Functions" in
the  "Related Pages" section of the doxygen documentation.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>