Added VScode support for Unix and Windows systems
This commit introduces a python script configureVScode.py that allows
FEAT to be configured and build by the VScode extension CMake Tools.
For a short guide, see the doxygen page linked on the main page.

This also introduces a few fixes to the build system regarding Clang
and its interaction with the pthread library (i.e. it is loaded
by default).
Also fixed a few other bugs in the configure setup and certain fixes
for windows specific problems.

If you only use the standard configure with gcc, you should not see any changes.